Sync "dbs" and "_users" with the next live node
The load induced by fully-connected mesh replication increases
quadratically with node count.  For large clusters with high rates of
database creation this ends up being significant.

This patch switches the topology to a ring.  Each node pushes to the
next live node in the ring.  Will deal with the correct action on
'nodeup' events in a separate commit.