Include target sequence in internal checkpoints
Saving both sequences allows us to be smarter about which sequence we
use in _changes feeds when replacing shards.  We can use a sequence on a
replacement copy that is wholly contained within the requested sequence
of the original copy and be assured that we don't skip any updates.
That maps to searching for a target seq less than the requested sequence
and then streaming from the corresponding source seq when we find it.

BugzID: 21973