Add helper to find source seq contained in target
This function is particularly useful for replacing shards in a _changes
feed, where we cannot afford to skip updates but we want to minimize the
number of updates that are duplicated in the feed.  When we know the
node / UUID / sequence of the original shard contributing to the feed
we can search the checkpoint history to find valid start sequences on
the replacment shard.

BugzID: 22698