Remove cached shards for non-existent dbs
When a database is deleted, a race can occur that will sometimes leave
its shard map present on one or more nodes in the cluster, which
generates a huge amount of logging.

When a request to mem3_shards:load_shards_from_db/2 results in
not_found, this sends a cast to mem3_shards on each node to remove the
shards for the specified DbName.